requirements.txt
matplotlib
pyspellchecker
pandas
googletrans==3.1.0a0